% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_edges.R
\name{compute_mi_std}
\alias{compute_mi_std}
\title{Compute the standardized mutual information between 2 variables}
\usage{
compute_mi_std(var1, var2, disc_method = "equalwidth", nbins = 20)
}
\arguments{
\item{var1}{vector containing data of variable 1}

\item{var2}{vector containing data of variable 2}

\item{disc_method}{discretization method}

\item{nbins}{number of bins}
}
\value{
(numeric value) standardized mutual information
}
\description{
Compute the standardized mutual information between 2 variables
}
\examples{
compute_mi_std(v1, v2)
}
