% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{cpdag2dag}
\alias{cpdag2dag}
\title{Extends CPDAG to DAG}
\usage{
cpdag2dag(cpdag)
}
\arguments{
\item{cpdag}{bnlearn object of the cpdag}
}
\value{
bnlearn object that is a DAG; if no extension if found, then original cpdag is returned
}
\description{
Extends CPDAG to DAG
}
\examples{
cpdag2dag(curr_cpdag)

}
