% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_edges.R
\name{creates_cycles}
\alias{creates_cycles}
\title{Check if adding edge induces cycle in graph}
\usage{
creates_cycles(adjmat, ch_ind, pa_ind, cpdag = T)
}
\arguments{
\item{adjmat}{adjacency matrix of graph}

\item{ch_ind}{name/index of child node}

\item{pa_ind}{name/index of parent node}

\item{cpdag}{evaluate as CPDAG or not (if not, evaluated as DAG)}
}
\value{
TRUE if edge induces cycle, FALSE otherwise
}
\description{
Check if adding edge induces cycle in graph
}
\examples{
creates_cycles(amat, "X1", "X2", cpdag=T)
}
