% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{DAGdatagen}
\alias{DAGdatagen}
\title{Generates data according to DAG structure}
\usage{
DAGdatagen(
  n = NULL,
  dag_amat = NULL,
  nodeN = nrow(dag_amat),
  labels = colnames(dag_amat),
  dagSparsityProb = NULL,
  data_fx = "inv",
  nonlinear.perct = 1,
  nonlinear.arcs = NULL,
  se = 1,
  ninv.method = "gp",
  inv.method = "random",
  error.distr = "gaussian",
  edge_fx_amat = NULL
)
}
\arguments{
\item{n}{number of samples}

\item{dag_amat}{adjacency matrix of DAG}

\item{nodeN}{number of nodes in (random) DAG}

\item{labels}{variable names}

\item{dagSparsityProb}{DAG sparsity parameter}

\item{data_fx}{options: linear, inv, ninv, nonlinear(inv and ninv)}

\item{nonlinear.perct}{percentage of edges generated by invertible, non-linear functions}

\item{nonlinear.arcs}{list of arcs to be generated by non-linear functions}

\item{se}{standard deviation value}

\item{ninv.method}{non-invertible function to use}

\item{inv.method}{invertible function to use}

\item{error.distr}{statistical distribution for error term}

\item{edge_fx_amat}{adjacency matrix containing pre-specified functions for each edge}
}
\value{
DAG object of class bnlearn
}
\description{
Given a DAG structure, the function generates according to the observed parent-child relations
Users may specify function types
If no DAG is provided, then function randomly generates a DAG
}
\examples{
amat <- rbind(c(0, 1, 1), c(0, 0, 1), c(0, 0, 0))
DAGdatagen(n=1000, dag_amat=amat, nodeN=rnow(amat))
}
