% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cpdag.R
\name{extend_to_dag}
\alias{extend_to_dag}
\title{Converts a PDAG to DAG}
\usage{
extend_to_dag(pdag, nodelabels)
}
\arguments{
\item{pdag}{bnlearn object of pdag}

\item{nodelabels}{vector of variables names}
}
\value{
adjacency matrix of dag if successful, adjacency matrix of pdag otherwise
}
\description{
Converts a PDAG to DAG
}
\examples{
extend_to_dag(example_pdag, c("X1", "X2", "X3"))
}
