% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_dag.R
\name{finalize_dag}
\alias{finalize_dag}
\title{(FOR PRACTICAL PURPOSES) STEP 4 OF ALGORITHM: convert PDAG to DAG}
\usage{
finalize_dag(
  data,
  curr_amat,
  nodelabels = colnames(data),
  alpha = 0.05,
  train_ratio = 0.5,
  ll_test_approach = NULL
)
}
\arguments{
\item{data}{observed data}

\item{curr_amat}{adjacency matrix of graph object}

\item{nodelabels}{vector containing variable names}

\item{alpha}{significance level for likelihood test}

\item{train_ratio}{percentage of data used for training models}

\item{ll_test_approach}{"2fold" for running the 2 fold test, other for 5-5 train test split approach}
}
\value{
adjacency matrix of DAG
}
\description{
In practice, the graph after pruning may contain undirected edges
Convert PDAG to DAG by applying likelihood test to determine true causal direction
}
\examples{
finalize_dag(data, curr_amat=adjmat, ll_test_approach='55')
}
