% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_test.R
\name{get_edge_dir}
\alias{get_edge_dir}
\title{Likelihood test for edge orientation (5/5 train test split)}
\usage{
get_edge_dir(m_XtoY, m_YtoX, m_x, m_y, test_data)
}
\arguments{
\item{m_XtoY}{model under X->Y; Y as function of PA_Y and X}

\item{m_YtoX}{model under Y->X; X as function of PA_X and Y}

\item{m_x}{model under X->Y; X as function of PA_X}

\item{m_y}{model under Y->X; Y as function of PA_Y}

\item{test_data}{data used to estimate log-lik}
}
\value{
list containing test results; 'pval_ll_diff' is the p-value, and 'llratio' indicates edge direction
}
\description{
With estimated regression models, compute the log-lik under each direction
Execute likelihood test to determine true causal direction
}
\examples{
get_edge_dir(gam_XtoY[[1]], gam_YtoX[[1]], gam_XtoY[[2]], gam_YtoX[[2]], test_data = test_data)
}
