% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_test.R
\name{get_edge_dir_2fold}
\alias{get_edge_dir_2fold}
\title{Likelihood test for edge orientation (2fold approach)}
\usage{
get_edge_dir_2fold(
  set_XtoY_1,
  set_YtoX_1,
  set_XtoY_2,
  set_YtoX_2,
  train_data1,
  train_data2
)
}
\arguments{
\item{set_XtoY_1}{model 1 under X->Y; Y as function of PA_Y and X, X as function of PA_X}

\item{set_YtoX_1}{model 1 under Y->X; X as function of PA_X and Y, Y as function of PA_Y}

\item{set_XtoY_2}{model 2 under X->Y; Y as function of PA_Y and X, X as function of PA_X}

\item{set_YtoX_2}{model 2 under Y->X; X as function of PA_X and Y, Y as function of PA_Y}

\item{train_data1}{data used to estimate log-lik for set 2}

\item{train_data2}{data used to estimate log-lik for set 1}
}
\value{
list containing test results; 'pval_ll_diff' is the p-value, and 'llratio' indicates edge direction
}
\description{
With estimated regression models, compute the log-lik under each direction
Execute likelihood test to determine true causal direction
Performs test twice, with train/test exchanging functions
}
\examples{
get_edge_dir_2fold(gam_XtoY_m1, gam_YtoX_m1, gam_XtoY_m2, gam_YtoX_m2,
train_data1=train_data, train_data2=test_data)
}
