% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_f1}
\alias{get_f1}
\title{Calculate results of learned DAG}
\usage{
get_f1(true_amat, temp_amat_list)
}
\arguments{
\item{true_amat}{ground truth graphical structure adjmat}

\item{temp_amat_list}{LIST of learned graphical structures adjmat}
}
\value{
vector/matrix containing results; each row corresponds to a learned dag
}
\description{
SHD, F1 score, true positives, false positives, false negatives, incorrectly oriented edges
}
\examples{
get_f1(true_dag, learned_dag_list)

}
