% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_edges.R
\name{get_gam_formula}
\alias{get_gam_formula}
\title{Builds GAM formula to be used in gam()}
\usage{
get_gam_formula(y, x, k, smooth_fx = "tp")
}
\arguments{
\item{y}{response variable name}

\item{x}{vector containing all covariate names}

\item{k}{number of basis functions for smoothing splines in GAM}

\item{smooth_fx}{basis function used}
}
\description{
Builds GAM formula to be used in gam()
}
\examples{
get_gam_formula("Y", c("X", "W", "Z"), k=3, smooth_fx="cr")
}
