% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_ji}
\alias{get_ji}
\title{Calculate jacard index (JI) of learned DAG}
\usage{
get_ji(true_amat, temp_amat)
}
\arguments{
\item{true_amat}{adjacency matrix of true DAG}

\item{temp_amat}{adjacency matrix of learned DAG}
}
\value{
return the jacard index
}
\description{
JI: percentage of correct edges among the union of edges in the estimated & true graphs
}
\examples{
get_ji(true_dag, learned_dag)

}
