% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_true_dag}
\alias{get_true_dag}
\title{Imports pre-saved/bnlearn graphs as bnlearn object}
\usage{
get_true_dag(network = "child", fixed_edges = TRUE)
}
\arguments{
\item{network}{name of network}

\item{fixed_edges}{true/statement on whether to keep certain edges fixed in cpdag}
}
\value{
vector/matrix containing results; each row corresponds to a learned dag
}
\description{
Includes self-defined graphs and graphs contained in .rda files
}
\examples{
get_true_dag('mildew')

}
