% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_edges.R
\name{ll_orient_test}
\alias{ll_orient_test}
\title{Perform likelihood ratio based orientation test on an undirected edge}
\usage{
ll_orient_test(
  data,
  train_ratio = 0.5,
  ll_test_approach = "2fold",
  adjmat,
  pa_ind,
  ch_ind,
  k_basis
)
}
\arguments{
\item{data}{observed data}

\item{train_ratio}{ratio ranging from 0-1 for train/test split on data}

\item{ll_test_approach}{"2fold" for running the 2 fold test, other for 5-5 train test split approach}

\item{adjmat}{adjacency matrix (of maximally oriented CPDAG)}

\item{pa_ind}{index/name of parent node}

\item{ch_ind}{index/name of child node}

\item{k_basis}{number of basis functions for GAM}
}
\value{
list containing likelihood ratio test result and statistics
}
\description{
Corresponds to the pairwise Additive Noise Models built for each direction in the paper
}
\examples{
ll_orient_test(data, ll_test_approach='55', adjmat=adjmat, pa_ind=2, ch_ind=5)
}
