% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_edges.R
\name{orient_edge}
\alias{orient_edge}
\title{STEP 2 OF ALGORITHM: orient undirected edges in the CPDAG}
\usage{
orient_edge(
  data,
  curr_amat,
  nodelabels = colnames(data),
  alpha = 0.05,
  train_ratio = 0.5,
  ll_test_approach = NULL
)
}
\arguments{
\item{data}{observed data}

\item{curr_amat}{adjacency matrix of graph object}

\item{nodelabels}{vector of variable names}

\item{alpha}{significance level for likelihood test}

\item{train_ratio}{percentage of data used for training data}

\item{ll_test_approach}{"2fold" for running the 2 fold test, other for 5-5 train test split approach}
}
\value{
adjacency matrix of updated PDAG
}
\description{
The procedure first categorizes edges by number of neighbors shared between nodes,
Computes an independence measure for each edge, which is used to sort edges,
And performs the likelihood test to determine the causal direction
}
\examples{
orient_edge(data, curr_amat=adjmat, ll_test_approach='55')
}
