% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_edges.R
\name{prune_dag}
\alias{prune_dag}
\title{STEP 3 OF ALGORITHM: Delete superfluous edges from PDAG to result in DAG}
\usage{
prune_dag(data, curr_amat, alpha = 1e-05, nodelabels = colnames(data))
}
\arguments{
\item{data}{observed data}

\item{curr_amat}{adjacency matrix of graph object}

\item{alpha}{significance level for covariate testing}

\item{nodelabels}{vector of node names}
}
\value{
adjacency matrix of (P)DAG
}
\description{
For each node, fit regression model using parents and neighboring nodes
Perform significance testing on covariates to remove edges
}
\examples{
prune_dag(data, curr_amat, alpha)
}
