% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sneo}
\alias{sneo}
\title{Nonlinear Causal Learning Algorithm by Likelihood Ratio Test}
\usage{
sneo(
  data,
  skel_alpha = 0.25,
  test_alpha = 0.05,
  covar_alpha = 1e-05,
  ll_test_approach = "5_5",
  train_ratio = 0.5,
  standardized = T
)
}
\arguments{
\item{data}{full dataset}

\item{skel_alpha}{sig. level for learning the initial skeleton}

\item{test_alpha}{sig. level for likelihood ratio test}

\item{covar_alpha}{sig. level for covariate testing}

\item{ll_test_approach}{variation of likelihood test to use; options are "2fold" and "5_5"}

\item{train_ratio}{percentage of data used for training dataset}

\item{standardized}{standardize data or not?}
}
\value{
adjacency matrix of learned DAG
}
\description{
Function for running the full algorithm
}
\examples{
sneo(data)
}
